int 64

+++ broker/model/submit_proposal_message.go
@@ -1,12 +1,12 @@
 package model
 
 type SubmitProposalMessage struct {
-	Height         int    `json:"height"`
+	Height         int64  `json:"height"`
 	TxHash         string `json:"tx_hash"`
-	MsgIndex       int    `json:"msg_index"`
+	MsgIndex       int64  `json:"msg_index"`
 	Proposer       string `json:"proposer"`
 	Messages       string `json:"messages"`
-	InitialDeposit int    `json:"initial_deposit"`
+	InitialDeposit int64  `json:"initial_deposit"`
 	Metadata       string `json:"metadata"`
 	Title          string `json:"title"`
 	Summary        string `json:"summary"`
